node: move parseRequiredLabels next to the label gate
