+++ internal/node/labels_gate.go
@@ -19,6 +19,7 @@ import (
 	"crypto/ed25519"
 	"fmt"
 	"sort"
+	"strings"
 	"time"
 
 	"github.com/google/sam/api"
@@ -27,6 +28,41 @@ import (
 	"github.com/libp2p/go-msgio"
 	"google.golang.org/protobuf/proto"
 )
+
+// parseRequiredLabels splits the X-Sam-Required-Labels header value
+// (comma-separated "key=value" pairs) into a label map; any malformed entry
+// rejects the whole request.
+func parseRequiredLabels(h string) (map[string]string, error) {
+	if h == "" {
+		return nil, nil
+	}
+	var out map[string]string
+	for _, part := range strings.Split(h, ",") {
+		part = strings.TrimSpace(part)
+		if part == "" {
+			continue
+		}
+		k, v, ok := strings.Cut(part, "=")
+		if !ok {
+			return nil, fmt.Errorf("invalid label %q: expected key=value", part)
+		}
+		k, v = strings.TrimSpace(k), strings.TrimSpace(v)
+		if err := api.ValidateLabelKey(k); err != nil {
+			return nil, err
+		}
+		if err := api.ValidateLabelValue(v); err != nil {
+			return nil, err
+		}
+		if out == nil {
+			out = make(map[string]string)
+		}
+		if _, exists := out[k]; exists {
+			return nil, fmt.Errorf("duplicate label key %q", k)
+		}
+		out[k] = v
+	}
+	return out, nil
+}
 
 // The label gate is the consumer-side enforcement point for label
 // requirements: gossip labels only rank providers, this gate verifies the
